Add support for periodic/reflected bc on bicubic_gsl
Because bicubic_gsl interpolators are natively implemented with natural
boundary conditions only, support is provided for periodic/reflected
conditions also. This is achieved by a simple extension on both edges of
the *second* variable domain, to where sampled data is copied following
periodic/reflected conditions. Data transposition for reordering
purposes was moved to the inside of the bicubic_gsl object.