Extend stellna to a 2nd-order near-axis expansion.
The class `equilibrium_stellna` is now being provided with its members
Covariant, Contravariant, Magnitude, and Jacobian overriden, along with
their respective derivatives. In this way, the need to explicity call
the metric components (which are correct only to 2nd-order terms in the
inverse aspect ratio) is avoided for lowering indices or other common
tasks. An error is issued if metric components are explicitly asked for.
The class `parser_stellna` was also deleted.