Recover previous stellna(qs) implementation
The initial stellna implementation was changed (commit 7a5f6cbba...) in
order to closely follow the properties exhibited by covariant and
contravariant components of the magnetic field in Boozer coordinates.
This move was motivated by much better momentum conservation when
following guiding centres in quasi-symmetric fields. However this came
with a significant cost attached: the transformation from Boozer to
cartesian coordinates is lost and the metric coefficients cannot be
defined. Therefore, the previous implementation is here recovered, even
if the truncation errors of the near-axis approach break the properties
of the Boozer representation to some extente and thus guiding-centre
momentum conservation is somewhat degraded. Also, stellna object names
(in metrics and fields) are appended with the suffix qs in order to
clearly signal that they may be used to represent quasi-symmetric field
only. The relevant documentation is updated and suitable references are
given.