Export the current git commit hash via `cmake`
Sometimes the version number is not sufficient to identify the specific
set of source-code files compiled for a given application. This is
particularly true because it needs to be updated manually in the file
`version.hh`. Now, `cmake` extracts automatically the current git commit
and exports it into the variable `gyronimo::git_commit_hash`.